requirement.txt
discord.py
pyautogui
PyGetWindow
audioop-lts
